@@ -34,38 +34,39 @@
 		"!dist/**/*.test.*"
 	],
 	"peerDependencies": {
-		"svelte": "^3.58.0 || ^4.0.0"
+		"svelte": "^5.0.0"
 	},
 	"repository": {
 		"type": "git",
 		"url": "git+https://github.com/bertybot/svelte-renderer.git"
 	},
 	"devDependencies": {
-		"@changesets/cli": "^2.26.2",
-		"@sveltejs/adapter-auto": "^2.1.0",
-		"@sveltejs/kit": "^1.27.2",
-		"@sveltejs/package": "^2.2.2",
-		"@testing-library/svelte": "^4.0.4",
-		"@typescript-eslint/eslint-plugin": "^5.62.0",
-		"@typescript-eslint/parser": "^5.62.0",
-		"eslint": "^8.52.0",
-		"eslint-config-prettier": "^8.10.0",
-		"eslint-plugin-svelte": "^2.34.0",
-		"happy-dom": "^8.9.0",
-		"jsdom": "^21.1.2",
-		"prettier": "^2.8.8",
-		"prettier-plugin-svelte": "^2.10.1",
-		"publint": "^0.2.5",
-		"svelte": "^4.2.2",
-		"svelte-check": "^3.5.2",
-		"tslib": "^2.6.2",
-		"typescript": "^5.2.2",
-		"vite": "^4.5.0",
-		"vitest": "^0.34.6"
+		"@changesets/cli": "^2.27.9",
+		"@sveltejs/adapter-auto": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/package": "^2.0.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"@testing-library/svelte": "^5.2.4",
+		"@typescript-eslint/eslint-plugin": "^8.11.0",
+		"@typescript-eslint/parser": "^8.11.0",
+		"eslint": "^9.13.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-svelte": "^2.46.0",
+		"happy-dom": "^15.7.4",
+		"jsdom": "^25.0.1",
+		"prettier": "^3.3.3",
+		"prettier-plugin-svelte": "^3.2.7",
+		"publint": "^0.2.11",
+		"svelte": "^5.0.5",
+		"svelte-check": "^4.0.5",
+		"tslib": "^2.8.0",
+		"typescript": "^5.6.3",
+		"vite": "^5.4.9",
+		"vitest": "^2.1.3"
 	},
 	"type": "module",
 	"dependencies": {
-		"@graphcms/rich-text-types": "^0.5.0",
+		"@graphcms/rich-text-types": "^0.5.1",
 		"es-escape-html": "^0.1.1"
 	},
 	"packageManager": "pnpm@8.8.0",
