@@ -1,3 +1,3 @@
 // Reexport your entry components here
 export { default as RichText } from './RichText.svelte';
-export * from './types';
+export * from './types.js';
