@@ -1,49 +1,55 @@
-import type { SvelteComponent } from 'svelte';
+import type { Snippet, SvelteComponent } from 'svelte';
 
 type AssetRendererType = {
-	application?: typeof SvelteComponent<object>;
-	audio?: typeof SvelteComponent<object>;
-	font?: typeof SvelteComponent<object>;
-	image?: typeof SvelteComponent<object>;
-	model?: typeof SvelteComponent<object>;
-	text?: typeof SvelteComponent<object>;
-	video?: typeof SvelteComponent<object>;
-	[key: string]: typeof SvelteComponent<object> | undefined;
+	application?: Snippet<[any]>;
+	audio?: Snippet<[any]>;
+	font?: Snippet<[any]>;
+	image?: Snippet<[any]>;
+	model?: Snippet<[any]>;
+	text?: Snippet<[any]>;
+	video?: Snippet<[any]>;
+	[key: string]: Snippet<[any]> | undefined;
 };
-export type NodeRendererType = {
-	a?: typeof SvelteComponent<object>;
-	class?: typeof SvelteComponent<object>;
-	img?: typeof SvelteComponent<object>;
-	video?: typeof SvelteComponent<object>;
-	iframe?: typeof SvelteComponent<object>;
-	h1?: typeof SvelteComponent<object>;
-	h2?: typeof SvelteComponent<object>;
-	h3?: typeof SvelteComponent<object>;
-	h4?: typeof SvelteComponent<object>;
-	h5?: typeof SvelteComponent<object>;
-	h6?: typeof SvelteComponent<object>;
-	p?: typeof SvelteComponent<object>;
-	ul?: typeof SvelteComponent<object>;
-	ol?: typeof SvelteComponent<object>;
-	li?: typeof SvelteComponent<object>;
-	list_item_child?: typeof SvelteComponent<object>;
-	table?: typeof SvelteComponent<object>;
-	table_head?: typeof SvelteComponent<object>;
-	table_body?: typeof SvelteComponent<object>;
-	table_row?: typeof SvelteComponent<object>;
-	table_cell?: typeof SvelteComponent<object>;
-	table_header_cell?: typeof SvelteComponent<object>;
-	blockquote?: typeof SvelteComponent<object>;
-	bold?: typeof SvelteComponent<object>;
-	italic?: typeof SvelteComponent<object>;
-	underline?: typeof SvelteComponent<object>;
-	code?: typeof SvelteComponent<object>;
-	code_block?: typeof SvelteComponent<object>;
+export type NodeRendererType = ElementRendererType & TextRendererType
+
+export type ElementRendererType = {
+	a?: Snippet<[any]>
+	class?: Snippet<[any]>
+	img?: Snippet<[any]>
+	image?: Snippet<[any]>
+	video?: Snippet<[any]>
+	iframe?: Snippet<[any]>
+	h1?: Snippet<[any]>
+	h2?: Snippet<[any]>
+	h3?: Snippet<[any]>
+	h4?: Snippet<[any]>
+	h5?: Snippet<[any]>
+	h6?: Snippet<[any]>
+	p?: Snippet<[any]>
+	ul?: Snippet<[any]>
+	ol?: Snippet<[any]>
+	li?: Snippet<[any]>
+	list_item_child?: Snippet<[any]>
+	table?: Snippet<[any]>
+	table_head?: Snippet<[any]>
+	table_body?: Snippet<[any]>
+	table_row?: Snippet<[any]>
+	table_cell?: Snippet<[any]>
+	table_header_cell?: Snippet<[any]>
+	blockquote?: Snippet<[any]>,
 	Asset?: AssetRendererType;
 	embed?: {
-		[key: string]: typeof SvelteComponent<object> | undefined;
+		[key: string]: Snippet<[any]>
 	};
 	link?: {
-		[key: string]: typeof SvelteComponent<object> | undefined;
+		[key: string]: Snippet<[any]>
 	};
-};
+}
+
+export type TextRendererType = {
+	bold?: Snippet;
+	italic?: Snippet;
+	underline?: Snippet;
+	code?: Snippet;
+	code_block?: Snippet<[any]>
+}
